auth_server/requirements.txt
flask==3.0.3
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gunicorn==22.0.0
